+++ src/hotspot/share/gc/shenandoah/shenandoahCollectionSet.hpp
@@ -26,6 +26,7 @@
 #define SHARE_GC_SHENANDOAH_SHENANDOAHCOLLECTIONSET_HPP
 
 #include "memory/allocation.hpp"
+#include "memory/virtualspace.hpp"
 #include "gc/shenandoah/shenandoahHeap.hpp"
 #include "gc/shenandoah/shenandoahHeapRegion.hpp"
 #include "gc/shenandoah/shenandoahPadding.hpp"
@@ -47,7 +48,7 @@ private:
   size_t                _region_count;
 
   shenandoah_padding(0);
-  volatile jint         _current_index;
+  volatile size_t       _current_index;
   shenandoah_padding(1);
 
 public:
